requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.13
django==4.2.4
Pillow==10.0.0